requirements.txt
streamlit
google-generativeai
python-docx
docxtpl
pdfplumber
pandas
streamlit-option-menu